+++ tests/testTS_TSI.py
@@ -594,7 +594,7 @@ class TestTS_TSI(unittest.TestCase):
 		}}
 		TestTS_TSI.sub, rsc = CREATE(tsURL, TestTS_TSI.originator, T.SUB, dct)
 		self.assertEqual(rsc, RC.created)
-		lastNotification = getLastNotification()
+		lastNotification = getLastNotification(wait = notificationDelay)
 		self.assertTrue(findXPath(lastNotification, 'm2m:sgn/vrq'))
 		self.assertTrue(findXPath(lastNotification, 'm2m:sgn/sur').endswith(findXPath(TestTS_TSI.sub, 'm2m:sub/ri')))
 
@@ -632,7 +632,7 @@ class TestTS_TSI(unittest.TestCase):
 			dgt += timeSeriesInterval
 
 			# Check notifications
-			lastNotification = getLastNotification(True)
+			lastNotification = getLastNotification(True, wait = notificationDelay)
 			if i % (maxMdn-2) == maxMdn-2-1:
 				self.assertIsNotNone(findXPath(lastNotification, 'm2m:sgn/nev/rep/m2m:tsn'), lastNotification)
 				self.assertEqual(len(findXPath(lastNotification, 'm2m:sgn/nev/rep/m2m:tsn/mdlt')), maxMdn-2, lastNotification)